feat: jsonb for sharing cont.

+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/query/DefaultQueryService.java
@@ -28,10 +28,7 @@ package org.hisp.dhis.query;
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import static com.google.common.base.Preconditions.checkNotNull;
-
-import java.util.List;
-
+import lombok.extern.slf4j.Slf4j;
 import org.hisp.dhis.common.IdentifiableObject;
 import org.hisp.dhis.fieldfilter.Defaults;
 import org.hisp.dhis.preheat.Preheat;
@@ -40,7 +37,9 @@ import org.hisp.dhis.query.planner.QueryPlanner;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
-import lombok.extern.slf4j.Slf4j;
+import java.util.List;
+
+import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
  * Default implementation of QueryService which works with IdObjects.
@@ -98,7 +97,7 @@ public class DefaultQueryService
     }
 
     @Override
-    public int count( Query query )
+    public long count( Query query )
     {
         Query cloned = Query.from( query );
 
@@ -145,7 +144,7 @@ public class DefaultQueryService
     // Helper methods
     //---------------------------------------------------------------------------------------------
 
-    private int countObjects( Query query )
+    private long countObjects( Query query )
     {
         List<? extends IdentifiableObject> objects;
         QueryPlan queryPlan = queryPlanner.planQuery( query );
